Update changelog and README for release 0.2.0
Added details for the new 0.2.0 release, including the change of the gem
name to "aigcm" and the release date. Expanded the README with a
detailed table of contents and context on usage options to improve
clarity for end users. This update ensures that users have a complete
understanding of the features and installation process of the gem.